Mis changes
1. Use '_' to replace ' ' for objects in testbed.py
2. Update the path in README
3. Add README to sso and lib folder

+++ samples/vsphere/common/vapiconnect.py
@@ -75,8 +75,9 @@ def logout(stub_config):
     """
     Delete session with vCenter.
     """
-    session_svc = Session(stub_config)
-    session_svc.delete()
+    if stub_config:
+        session_svc = Session(stub_config)
+        session_svc.delete()
 
 
 def create_unverified_session(session, suppress_warning):
